requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
jinja2>=3.0.0
python-dotenv>=1.0.0
requests>=2.28.0
